rerfactor: 도서 산간 지역 배송비 선택 상태 추가, 배송비 관련 상태를 shipping.ts로 분리
