+++ src/recoil/order.ts
@@ -56,16 +56,3 @@ export const numberOfTypesSelector = selector({
     return numberOfTypes;
   },
 });
-
-/**
- * 선택된 제품의 배송비. 주문 금액이 100,000원 이상이면 0원
- */
-export const shippingFeeSelector = selector({
-  key: 'shippingFeeSelector',
-  get: ({ get }) => {
-    const totalPrice = get(totalPriceSelector);
-
-    if (totalPrice >= 100_000) return 0;
-    return 3000;
-  },
-});
